+++ com.unity.formats.fbx/Editor/FbxExporter.cs
@@ -174,9 +174,10 @@ namespace UnityEditor.Formats.Fbx.Exporter
         Dictionary<int, FbxSurfaceMaterial> MaterialMap = new Dictionary<int, FbxSurfaceMaterial> ();
 
         /// <summary>
-        /// Map texture filename name to FBX texture object
+        /// Map texture properties to FBX texture object
         /// </summary>
-        Dictionary<string, FbxTexture> TextureMap = new Dictionary<string, FbxTexture> ();
+        Dictionary<(Texture unityTexture, Vector2 offset, Vector2 scale, TextureWrapMode wrapModeU, TextureWrapMode wrapModeV), FbxFileTexture> TextureMap = 
+            new Dictionary<(Texture unityTexture, Vector2 offset, Vector2 scale, TextureWrapMode wrapModeU, TextureWrapMode wrapModeV), FbxFileTexture>();
 
         /// <summary>
         /// Map a Unity mesh to an fbx node (for preserving instances) 
@@ -634,20 +635,44 @@ namespace UnityEditor.Formats.Fbx.Exporter
                 return false;
             }
 
+            var offset = unityMaterial.GetTextureOffset(unityPropName);
+            var scale = unityMaterial.GetTextureScale(unityPropName);
+            var wrapModeU = unityTexture.wrapModeU;
+            var wrapModeV = unityTexture.wrapModeV;
+
+            var tuple = (unityTexture, offset, scale, wrapModeU, wrapModeV);
+
+            FbxFileTexture fbxTexture;
             // Find or create an fbx texture and link it up to the fbx material.
-            if (!TextureMap.ContainsKey (textureSourceFullPath)) {
+            if (!TextureMap.TryGetValue(tuple, out fbxTexture))
+            {
                 var textureName = GetUniqueTextureName(fbxPropName + "_Texture");
-                var fbxTexture = FbxFileTexture.Create (fbxMaterial, textureName);
+                fbxTexture = FbxFileTexture.Create (fbxMaterial, textureName);
                 fbxTexture.SetFileName (textureSourceFullPath);
                 fbxTexture.SetTextureUse (FbxTexture.ETextureUse.eStandard);
                 fbxTexture.SetMappingType (FbxTexture.EMappingType.eUV);
-                TextureMap.Add (textureSourceFullPath, fbxTexture);
+                fbxTexture.SetScale(scale.x, scale.y);
+                fbxTexture.SetTranslation(offset.x, offset.y);
+                fbxTexture.SetWrapMode(GetWrapModeFromUnityWrapMode(wrapModeU, unityMaterial.name, unityPropName),
+                                       GetWrapModeFromUnityWrapMode(wrapModeV, unityMaterial.name, unityPropName));
+                TextureMap.Add (tuple, fbxTexture);
             }
-            TextureMap [textureSourceFullPath].ConnectDstProperty (fbxMaterialProperty);
+            fbxTexture.ConnectDstProperty(fbxMaterialProperty);
 
             return true;
         }
 
+        private FbxTexture.EWrapMode GetWrapModeFromUnityWrapMode(TextureWrapMode wrapMode, string materialName, string textureName)
+        {
+            switch (wrapMode)
+            {
+                case TextureWrapMode.Clamp: return FbxTexture.EWrapMode.eClamp;
+                case TextureWrapMode.Repeat: return FbxTexture.EWrapMode.eRepeat;
+                default:
+                    Debug.LogWarning($"Texture {textureName} on material {materialName} uses wrap mode {wrapMode} which is not supported by FBX. Will be exported as \"Repeat\".");
+                    return FbxTexture.EWrapMode.eRepeat;
+            }
+        }
         /// <summary>
         /// Get the color of a material, or grey if we can't find it.
         /// </summary>